Automatically handle DeveloperID certificates.
Rather than keying off the "ACTION" environment variable, we can key off the fact that it is a DeveloperID certificate - which causes the binary to be signed with a different Designated Requirement.  Previously, the Developer ID designated requirement was used only when ACTION was set to "install".  This change keys off of the certificate CN, and if it is a Developer ID certificate, it will generate the strings appropriate for that signature.